+++ src/util.cc
@@ -302,7 +302,7 @@ char* get_word(char*& s)
 	return w;
 	}
 
-void get_word(int length, const char* s, int& pwlen, const char*& pw)
+void get_word(uint64_t length, const char* s, int& pwlen, const char*& pw)
 	{
 	pw = s;
